Preparing for new netplay UI
