--- src/netplay/stats.rs
@@ -1,38 +1,0 @@
-use std::{
-    collections::VecDeque,
-    time::{Duration, Instant},
-};
-
-use ggrs::NetworkStats;
-
-pub const STATS_HISTORY: usize = 100;
-
-pub struct NetplayStat {
-    pub stat: NetworkStats,
-    pub duration: Duration,
-}
-pub struct NetplayStats {
-    stats: VecDeque<NetplayStat>,
-    start_time: Instant,
-}
-
-impl NetplayStats {
-    pub fn new() -> Self {
-        Self {
-            start_time: Instant::now(),
-            stats: VecDeque::with_capacity(STATS_HISTORY),
-        }
-    }
-
-    pub fn get_ping(&self) -> &VecDeque<NetplayStat> {
-        &self.stats
-    }
-
-    pub fn push_stats(&mut self, stat: NetworkStats) {
-        let duration = Instant::now().duration_since(self.start_time);
-        self.stats.push_back(NetplayStat { duration, stat });
-        if self.stats.len() == STATS_HISTORY {
-            self.stats.pop_front();
-        }
-    }
-}
