+++ src/settings/gui.rs
@@ -3,15 +3,12 @@ use std::{
     time::{Duration, Instant},
 };
 
-use egui::{
-    Align2, Color32, Context, CursorIcon, FontId, Id, Margin, Response, RichText, Sense, TextStyle,
-    Ui, Vec2, Widget, WidgetInfo, WidgetText, WidgetType, Window,
-};
-use winit::dpi::LogicalSize;
+use egui::{Color32, Context, FontId, Margin, Response, RichText, Ui, Widget};
 
 use crate::{
     audio::gui::AudioGui,
     emulation::gui::EmulatorGui,
+    gui::{centered_window, MenuButton},
     input::{
         buttons::GamepadButton, gamepad::GamepadEvent, gui::InputsGui, keys::KeyCode, KeyEvent,
     },
@@ -42,71 +39,6 @@ pub trait GuiComponent {
         None
     }
     fn handle_event(&mut self, _gui_event: &GuiEvent) {}
-}
-
-pub struct MainMenuButton {
-    text: WidgetText,
-    sense: Sense,
-}
-impl MainMenuButton {
-    const ACTIVE_COLOR: Color32 = Color32::WHITE;
-    const UNACTIVE_COLOR: Color32 = Color32::from_rgb(96, 96, 96);
-    const LAST_FOCUS_KEY: &'static str = "LAST_FOCUS";
-
-    fn new(text: impl Into<String>) -> Self {
-        Self {
-            text: RichText::new(text)
-                .color(Color32::PLACEHOLDER)
-                .strong()
-                .font(FontId::monospace(30.0))
-                .into(),
-            sense: Sense::click(),
-        }
-    }
-}
-impl Widget for MainMenuButton {
-    fn ui(self, ui: &mut Ui) -> egui::Response {
-        let mut desired_size = Vec2::ZERO;
-        let galley =
-            self.text
-                .into_galley(ui, Some(false), ui.available_width(), TextStyle::Button);
-
-        desired_size.x += galley.size().x;
-        desired_size.y = desired_size.y.max(galley.size().y);
-        let (rect, mut response) = ui.allocate_at_least(desired_size, self.sense);
-        response.widget_info(|| WidgetInfo::labeled(WidgetType::Button, galley.text()));
-
-        ui.memory_mut(|m| {
-            if response.hovered() {
-                m.request_focus(response.id);
-            } else {
-                let last_focused_data_key = Id::new(MainMenuButton::LAST_FOCUS_KEY);
-                if let Some(current_focus) = m.focus() {
-                    m.data.insert_temp(last_focused_data_key, current_focus);
-                } else if let Some(last_focused_id) = m.data.get_temp(last_focused_data_key) {
-                    m.request_focus(last_focused_id);
-                } else {
-                    m.request_focus(response.id);
-                }
-            }
-        });
-
-        if ui.is_rect_visible(rect) {
-            let text_pos = ui.layout().align_size_within_rect(galley.size(), rect).min;
-            response = response.on_hover_cursor(CursorIcon::PointingHand);
-
-            ui.painter().galley(
-                text_pos,
-                galley,
-                if response.has_focus() {
-                    Self::ACTIVE_COLOR
-                } else {
-                    Self::UNACTIVE_COLOR
-                },
-            );
-        }
-        response
-    }
 }
 
 #[derive(Debug)]
@@ -145,28 +77,11 @@ impl SettingsGui {
     }
 
     fn menu_item_ui(ui: &mut Ui, text: impl Into<String>) -> Response {
-        let res = ui.vertical_centered(|ui| MainMenuButton::new(text).ui(ui));
+        let res = ui.vertical_centered(|ui| MenuButton::new(text).ui(ui));
         ui.end_row();
         res.inner
     }
 
-    fn main_window(
-        window: &Arc<winit::window::Window>,
-        ctx: &Context,
-        title: Option<&str>,
-        content: impl FnOnce(&mut Ui),
-    ) {
-        let size: LogicalSize<f32> = window.inner_size().to_logical(window.scale_factor());
-
-        Window::new(title.unwrap_or(""))
-            .title_bar(title.is_some())
-            .collapsible(false)
-            .resizable(false)
-            .movable(false)
-            .pivot(Align2::CENTER_CENTER)
-            .fixed_pos([size.width / 2.0, size.height / 2.0])
-            .show(ctx, content);
-    }
     pub fn ui(
         &mut self,
         ctx: &Context,
@@ -177,7 +92,7 @@ impl SettingsGui {
         if self.visible {
             match &mut self.state {
                 MainMenuState::Main => {
-                    Self::main_window(&self.window, ctx, None, |ui| {
+                    centered_window(&self.window, ctx, None, |ui| {
                         ui.add_space(20.0);
                         egui::Grid::new("main_menu_grid")
                             .num_columns(1)
@@ -205,7 +120,7 @@ impl SettingsGui {
                     });
                 }
                 MainMenuState::Settings => {
-                    Self::main_window(&self.window, ctx, Some("Settings"), |ui| {
+                    centered_window(&self.window, ctx, Some("Settings"), |ui| {
                         if let Some(name) = audio_gui.name() {
                             ui.vertical_centered(|ui| {
                                 ui.heading(name);
@@ -231,7 +146,7 @@ impl SettingsGui {
                 }
                 MainMenuState::Netplay => {
                     if let Some(name) = emulator_gui.name() {
-                        Self::main_window(&self.window, ctx, Some(&name), |ui| {
+                        centered_window(&self.window, ctx, Some(&name), |ui| {
                             emulator_gui.ui(ui);
                         });
                     }
@@ -271,7 +186,9 @@ impl SettingsGui {
     pub fn handle_event(
         &mut self,
         gui_event: &GuiEvent,
-        gui_components: &mut [&mut dyn GuiComponent],
+        audio_gui: &mut AudioGui,
+        inputs_gui: &mut InputsGui,
+        emulator_gui: &mut EmulatorGui,
     ) {
         match gui_event {
             GuiEvent::Gamepad(crate::input::gamepad::GamepadEvent::ButtonDown {
@@ -293,6 +210,9 @@ impl SettingsGui {
                 }
             }
             _ => {
+                let gui_components: &mut [&mut dyn GuiComponent] =
+                    &mut [audio_gui, inputs_gui, emulator_gui];
+
                 for gui in gui_components {
                     gui.handle_event(gui_event);
                 }
